some moving stuff & ignore failing tests
